+++ src/websocket.cpp
@@ -32,14 +32,26 @@ static std::shared_ptr<boost::asio::ssl::context> CreateSSLContext(
                    boost::asio::ssl::context::no_sslv3 |
                    boost::asio::ssl::context::single_dh_use);
   if (client_cert) {
+    boost::system::error_code ec;
     ctx->use_certificate(boost::asio::buffer(*client_cert),
-                         boost::asio::ssl::context_base::file_format::pem);
-    RTC_LOG(LS_INFO) << "client_cert=" << *client_cert;
+                         boost::asio::ssl::context_base::file_format::pem, ec);
+    if (ec) {
+      RTC_LOG(LS_WARNING) << "client_cert is set, but use_certificate failed: "
+                          << ec.message();
+    } else {
+      RTC_LOG(LS_INFO) << "client_cert is set";
+    }
   }
   if (client_key) {
+    boost::system::error_code ec;
     ctx->use_private_key(boost::asio::buffer(*client_key),
-                         boost::asio::ssl::context_base::file_format::pem);
-    RTC_LOG(LS_INFO) << "client_key=" << *client_key;
+                         boost::asio::ssl::context_base::file_format::pem, ec);
+    if (ec) {
+      RTC_LOG(LS_WARNING) << "client_key is set, but use_private_key failed: "
+                          << ec.message();
+    } else {
+      RTC_LOG(LS_INFO) << "client_key is set";
+    }
   }
   return ctx;
 }
@@ -180,7 +192,7 @@ void Websocket::Connect(const std::string& url, connect_callback_t on_connect) {
 
   // ヘッダーの設定
   auto set_headers = [this](boost::beast::websocket::request_type& req) {
-    if (user_agent_ != boost::none) {
+    if (user_agent_ != std::nullopt) {
       req.set(boost::beast::http::field::user_agent, *user_agent_);
     }
   };