+++ CHANGES.md
@@ -11,17 +11,23 @@
 
 ## develop
 
+- [CHANGE] `boost::optional` を利用していた部分を全て `std::optional` に変更
+  - @melpon
 - [UPDATE] VPL を v2.13.0 に上げる
   - @voluntas
-- [UPDATE] CMake を 3.30.4 にあげる
+- [UPDATE] CMake を 3.30.5 にあげる
   - @voluntas
 - [UPDATE] Boost を 1.86.0 にあげる
   - @voluntas
-- [UPDATE] libwebrtc を m129.6668.1.0 にあげる
+- [UPDATE] libwebrtc を m131.6778.3.1 にあげる
   - H.265 Patch の修正に伴い、C++ SDK の H.265 に関する設定を変更
   - examples と test に `rtc::CreateRandomString` のヘッダを追加
   - `SetRtpTimestamp` を `frame.timestamp` から `frame.rtp_timestamp` に変更
-  - @tnoho @torikizi
+  - `scalable_track_source.h` と `scalable_track_source.cpp` の `absl::optional` を `std::optional` に変更
+  - `nvcodec_video_decoder.h` と `vpl_video_decoder.h` に `#include <optional>` を追加
+  - `nvcodec_video_decoder.cpp` と `vpl_video_decoder.cpp` の `absl::nullopt` を `std::nullopt` に変更
+  - Android の test アプリの cmake バージョンを VERSION と合わせる
+  - @tnoho @torikizi @melpon
 - [UPDATE] Xcode のバージョンを 15.4 にあげる
   - @tnoho
 - [UPDATE] SDL を 2.30.8 に上げる
@@ -30,6 +36,8 @@
   - @torikizi
 - [UPDATE] ASMJIT_VERSION を上げる
   - @torikizi
+- [UPDATE] OpenH264 を v2.5.0 に上げる
+  - @voluntas
 - [ADD] シグナリングメッセージを取得できるよう OnSignalingMessage を SoraSignalingObserver に追加する
   - @tnoho
 - [ADD] Intel VPL で AV1 エンコーダを動くようにする
@@ -42,6 +50,18 @@
   - @melpon
 - [ADD] WebSocket の Close を取得できるよう SendOnWsClose を SoraSignalingObserver に追加する
   - @tnoho
+- [ADD] DataChannel のみの接続で type: close がやってきた場合に正しく切断されるようにする
+  - @melpon
+- [ADD] SoraSignalingConfig に audio_opus_params を追加
+  - @melpon
+- [ADD] SoraSignalingConfig::DataChennel に header を追加
+  - @melpon
+- [ADD] SoraSignalingConfig::ForwardingFilter に name と priority を追加
+  - @melpon
+- [ADD] SoraSignalingConfig に forwarding_filters を追加
+  - @melpon
+- [ADD] scaleResolutionDownTo に対応する
+  - @melpon
 - [FIX] HTTP Proxy 利用時の Websocket 初期化で insecure_ メンバ変数が初期化されていなかったのを修正
   - @melpon
 - [FIX] SoraSignalingConfig の client_cert と client_key に渡す必要がある値を、ファイルパスからファイルの内容に修正
@@ -54,6 +74,8 @@
 
 ### misc
 
+- [UPDATE] CATCH2 を v3.7.1 にあげる
+  - @voluntas
 - [UPDATE] SDL2 を 2.30.7 にあげる
   - @voluntas
 - [ADD] sumomo に証明書に関するオプションを追加する